requirements.txt
APScheduler==3.7.0
click==7.1.2
Flask==1.1.2
Flask-APScheduler==1.11.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyserial==3.5
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
tzlocal==2.1
Werkzeug==1.0.1
